feat: Refactor weighted module rendering on module page,
The code changes refactor the rendering of weighted modules on the module page. The `Page` component now uses the `useDelegateStore` hook to retrieve the `delegatedModules` and filter out modules with a percentage greater than 0. The filtered modules are then rendered using the `ModuleCard` component. This change improves the display of weighted modules on the page. This also allows us to update the delegation view on frontend based on what the user has submitted to the db.